buildLangPrefix retornado a languagePickerUtils
