+++ src/domain/translationLink.ts
@@ -1,6 +1,4 @@
 import type { UILanguages } from '@i18n/ui'
-import { defaultLang } from '@i18n/ui'
-import { showDefaultLang } from '@i18n/config'
 
 export interface TranslationLink {
   /** URL destino del link (puede ser string vacío si no está disponible) */
@@ -25,12 +23,4 @@ export function availableLink(href: string, lang: UILanguages): TranslationLink 
  */
 export function missingLink(lang: UILanguages): TranslationLink {
   return { href: '', isAvailable: false, disabledReason: 'missing', locale: lang }
-}
-
-/**
- * Calcula el prefijo de ruta localizado según el idioma.
- * Ejemplo: '/en' para inglés, '' para idioma por defecto (si showDefaultLang es false).
- */
-export function buildLangPrefix(targetLang: UILanguages): string {
-  return targetLang === defaultLang && !showDefaultLang ? "" : `/${targetLang}`
 }
