+++ src/utils/staticPathsBuilder.ts
@@ -19,7 +19,8 @@ import type { AvailableLocales } from '@domain/translation'
 import { type PostEntry, type ExperienceLikeEntry } from '@domain/post'
 import { buildTagLocaleMap, getAvailableLocaleEntriesFromMap, buildLocaleEntryMap } from './translationHelpers'
 import { buildDetailLinks } from '@i18n/languagePickerUtils'
-import { availableLink, missingLink, buildLangPrefix } from '@domain/translationLink'
+import { availableLink, missingLink } from '@domain/translationLink'
+import { buildLangPrefix } from '@i18n/languagePickerUtils'
 import type { TranslationLink } from '@domain/translationLink'
 import { tagTranslations } from '@domain/tags'
 import type { CollectionEntry, CollectionKey } from 'astro:content'
